+++ apps/server/src/modules/access-logs/metrics/types.ts
@@ -26,6 +26,8 @@ export const AccessLogMetricsSchema = t.Object({
 	currentStates: t.Object({
 		rps: t.Union([t.Number(), t.Any()]),
 		statusCodes,
+		maxMemory: t.Number(),
+		usedMemory: t.Number(),
 	}),
 	globalStates: t.Object({
 		bytes: t.Number(),
@@ -35,6 +37,7 @@ export const AccessLogMetricsSchema = t.Object({
 		hitRatePercent: t.Number(),
 		successRatePercent: t.Number(),
 		contentTypes: contentTypeStats,
+		rpsHistory: t.Array(t.Number()),
 	}),
 	users: t.Array(
 		t.Object({