+++ apps/web/src/components/access-metric/masonry-wall/BCardRPS.vue
@@ -1,11 +1,17 @@
 <script setup lang="ts">
+import BRpsHistory from "@/components/access-metric/BRpsHistory.vue";
 import BCardMetric from "@/components/BCardMetric.vue";
 import { diffDate } from "@/utils/date.ts";
 
-const { rps, diffTime, time } = defineProps<{
+const {
+	rps,
+	diffTime,
+	time = [0, 60000],
+} = defineProps<{
 	rps: number;
 	diffTime: number;
 	time?: [number, number];
+	rpsHistory: number[];
 }>();
 </script>
 
@@ -16,6 +22,8 @@ const { rps, diffTime, time } = defineProps<{
     <span class="text-lg">
         {{ (rps * diffTime).toFixed(0) }} in {{ diffDate(time) }}
     </span>
+
+    <BRpsHistory :rps-history="rpsHistory" />
 
     <template #name>
       {{ $t('rpsTitle') }}